requirements.txt
torch==2.4.1+cu128
torchvision==0.19.1+cu128
torchaudio==2.4.1+cu128
opencv-python==4.10.0
albumentations==1.4.14
wandb==0.17.5
fastapi==0.112.0
uvicorn==0.30.3
ultralytics==8.2.74
h5py==3.11.0
pyyaml==6.0.1
twilio==9.2.3
tqdm==4.66.5
numpy==1.26.4
scikit-learn==1.5.1